fix: remaining issues
