+++ app/javascript/pages/account/components/NameFieldset.tsx
@@ -28,6 +28,15 @@ export const handleNameServerErrors = (
     : [name, { message: "name:server:generic", shouldFocus: true }]
 }
 
+const nameValidation = (name: "firstName" | "lastName", value: string) => {
+  if (value.includes("www") || value.includes("http") || value.includes(".")) {
+    return `name:server:generic`
+  }
+  if (value.trim() === "") {
+    return `name:${name}`
+  }
+}
+
 export const nameFieldsetErrors: ErrorMessages = {
   "name:firstName": {
     default: "error.account.firstName",
@@ -68,7 +77,7 @@ const NameFieldset = ({
         defaultValue={defaultFirstName ?? null}
         validation={{
           required: "name:firstName",
-          validate: (data: string) => data.trim() !== "" || "name:firstName",
+          validate: (data: string) => nameValidation("firstName", data),
         }}
         inputProps={{ required: true }}
         onChange={onChange}
@@ -95,7 +104,7 @@ const NameFieldset = ({
         register={register}
         validation={{
           required: "name:lastName",
-          validate: (data: string) => data.trim() !== "" || "name:firstName",
+          validate: (data: string) => nameValidation("lastName", data),
         }}
         onChange={onChange}
         inputProps={{ required: true }}
