+++ app/javascript/pages/account/create-account.scss
@@ -1,0 +1,3 @@
+.error-message {
+  letter-spacing: normal;
+}
